Organize zoned skipping indexes and aggregates
