+++ vortex-file/tests/bloom_skip_index.rs
@@ -38,9 +38,9 @@ use vortex_file::WriteOptionsSessionExt;
 use vortex_file::WriteStrategyBuilder;
 use vortex_io::session::RuntimeSession;
 use vortex_layout::LayoutStrategy;
-use vortex_layout::layouts::zoned::skip_index::BloomOptions;
-use vortex_layout::layouts::zoned::skip_index::BloomSkipIndex;
 use vortex_layout::layouts::zoned::skip_index::SkipIndex;
+use vortex_layout::layouts::zoned::skip_index::bloom::BloomOptions;
+use vortex_layout::layouts::zoned::skip_index::bloom::BloomSkipIndex;
 use vortex_layout::layouts::zoned::writer::ZonedLayoutOptions;
 use vortex_layout::segments::SegmentFuture;
 use vortex_layout::segments::SegmentId;
