+++ vortex-layout/src/layouts/zoned/aggregates/min_max.rs
@@ -1,0 +1,70 @@
+//! Min/max aggregate selection for zoned layouts.
+
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use vortex_array::aggregate_fn::AggregateFnRef;
+use vortex_array::aggregate_fn::AggregateFnVTableExt;
+use vortex_array::aggregate_fn::NumericalAggregateOpts;
+use vortex_array::aggregate_fn::fns::bounded_max::BoundedMax;
+use vortex_array::aggregate_fn::fns::bounded_max::BoundedMaxOptions;
+use vortex_array::aggregate_fn::fns::bounded_min::BoundedMin;
+use vortex_array::aggregate_fn::fns::bounded_min::BoundedMinOptions;
+use vortex_array::aggregate_fn::fns::max::Max;
+use vortex_array::aggregate_fn::fns::min::Min;
+use vortex_array::dtype::DType;
+
+use super::super::schema::default_bounded_stat_max_bytes;
+
+pub(super) fn min_max_aggregate_fns(dtype: &DType) -> [AggregateFnRef; 2] {
+    match dtype {
+        DType::Utf8(_) | DType::Binary(_) => [
+            BoundedMax.bind(BoundedMaxOptions {
+                max_bytes: default_bounded_stat_max_bytes(),
+            }),
+            BoundedMin.bind(BoundedMinOptions {
+                max_bytes: default_bounded_stat_max_bytes(),
+            }),
+        ],
+        _ => [
+            Max.bind(NumericalAggregateOpts::skip_nans()),
+            Min.bind(NumericalAggregateOpts::skip_nans()),
+        ],
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use vortex_array::aggregate_fn::fns::bounded_max::BoundedMax;
+    use vortex_array::aggregate_fn::fns::bounded_min::BoundedMin;
+    use vortex_array::aggregate_fn::fns::max::Max;
+    use vortex_array::aggregate_fn::fns::min::Min;
+    use vortex_array::dtype::DType;
+    use vortex_array::dtype::Nullability;
+    use vortex_array::dtype::PType;
+
+    use super::default_bounded_stat_max_bytes;
+    use super::min_max_aggregate_fns;
+
+    #[test]
+    fn variable_length_min_max_are_bounded() {
+        let aggregate_fns = min_max_aggregate_fns(&DType::Utf8(Nullability::NonNullable));
+
+        assert_eq!(
+            aggregate_fns[0].as_::<BoundedMax>().max_bytes,
+            default_bounded_stat_max_bytes()
+        );
+        assert_eq!(
+            aggregate_fns[1].as_::<BoundedMin>().max_bytes,
+            default_bounded_stat_max_bytes()
+        );
+    }
+
+    #[test]
+    fn fixed_width_min_max_are_exact() {
+        let aggregate_fns = min_max_aggregate_fns(&PType::I32.into());
+
+        assert!(aggregate_fns[0].is::<Max>());
+        assert!(aggregate_fns[1].is::<Min>());
+    }
+}
