+++ vortex-layout/src/layouts/zoned/mod.rs
@@ -11,6 +11,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+mod aggregates;
 mod builder;
 mod pruning;
 mod reader;
